Add support to block a specific user
Welcome to the internets where not everyone understand what
they are doing.

Adds the Account.blocked bit. This bit only remove the ability
to reply to a message or to use the Publish command. The user
can still upload an issue.

Add a form and a link on each user page. Only show the link for
administrators.

Review: https://codereview.appspot.com/7057066/