Actions in garden-o-matic are only available for items above the fold
Currently in garden-o-matic it is not possible to examine or rebaseline
tests that are below the fold, regardless of scroll position. When hovering
lines below the fold no action buttons appears.

While I assume the bug is trivial to fix I'm unable to find the code that
applies the active class to the list items or the mouse event handlers
responsible. This fix instead changes the list items to have relative
position which, I speculate, makes the scroll position irrelevant for the
calculation.

R=dglazkov@chromium.org
BUG=236957

Review URL: https://codereview.chromium.org/14683003

git-svn-id: svn://svn.chromium.org/blink/trunk@149496 bbb929c8-8fbe-4397-9dbb-9b2b20218538