chromium-status: localize timestamps
While UTC is great for robots, it sucks for humans.  So let's translate
things to the user's local timezone.

Unfortunately, appengine does not make this easy (which is partially due
to HTTP clients not communicating it easily), so instead, let's do it all
on the fly client side.

The browser knows the exact timezone that is active and javascript makes
it easy to convert from UTC to it.  There is a slight refresh delay as
the code kicks in, but it's not a big deal.

BUG=chromium:314501
TEST=ran an appengine instance and checked the timestamps get converted correctly
TEST=`./tests/main_test.py` passes

Review URL: https://codereview.chromium.org/63603008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/chromium-status@239256 0039d316-1c4b-4281-b951-d872f2087c98