Use preferred host name when sending reminder email.
When the user sends an l-g-t-m via email, we ignore that and send by a
reminder email with a link to the website.  But, we don't know exactly
which domain name to use in that link because app aliases go through
GAFYD rather than anything in AppEngine, and the request with the email
body hits our app at a canonical domain.

This CL makes the link use the preferred domain name defined in
settings.py.

Code review URL: https://codereview.appspot.com/84470048/