Give chromium-status support for a Git-only world
The chromium-status /revisions and /lkgr endpoints have support for Git,
but not for Git-only. They are reliant on sorting SVN revisions, and treat
Git hashes as an afterthought. This CL upgrades Git hashes to first-class
citizens and prepares to remove SVN support entirely.

It also performs some cleanup to remove the deprecated Status and Failed Tests
fields.

BUG=261731
R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/178683007