codereview: preserve first mail subject for entire review thread
Otherwise, if someone fixes a typo in an issue subject,
it splits the review into a separate mail thread.
People replying to the original mail end up in the first thread
and people replying on the web page end up in a second
(or third or fourth...) thread.

Using the original subject means we're stuck with the typos
and mistakes in the original subject (as it existed when the first
mail was sent) but it keeps the conversation in one thread.

This is by far the biggest pain point I have when using Rietveld.

Note that the subject can still be changed at will before sending
the first message. It's only the act of sending a mail that freezes
the mail subject.

Note also that the issue subject, as shown on the web page,
is unaffected: it can still change and always shows the latest one.

LGTM=jrobbins
R=jrobbins
https://codereview.appspot.com/102650046