Simplify failing builder lists.
-Get rid of the "BUILDING" category. In order to make this work,
we'll need to plumb through a list of builders for each testtype/waterfall
combination. While we have that information for the flakiness dashboard,
it just doesn't seem worth it for this feature.
-Stop trying to do complicated things with the builder name
and just show the raw builder name. It takes up more space,
but it scales better to the full list of chromium builders and
is less complexity for sheriffs to understand.
-Once we've done the above, stop plumbing builderLatestRevisions
everywhere. Technically it's dead code now, but Levi's patch
in progress is using it, so I don't feel compelled to delete it.
-Finally, we can now remove config.js entirely.

NOTRY=true

Review URL: https://codereview.chromium.org/409063002

git-svn-id: svn://svn.chromium.org/blink/trunk@178625 bbb929c8-8fbe-4397-9dbb-9b2b20218538