GS index.html now presents full set of objects
Updated page to repeatedly query 'nextPageToken' links to retrieve the full set
of storage objects.

Because of this slowness, the 'sort' link has been removed, as clicking a sort
link causes the page to be completely reloaded and thus repeats the long query.

This is slow with large buckets, and rendering doesn't happen until after the
full query set has completed. If this page is actually a first-class capability,
we need to spend the effort to make this a single stateful paginated page.

BUG=384681
TEST=live
R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/433533002