Import auto-sheriff.appspot.com feeder in to infra
dcommitting to unblock ojan.  This has been waiting to land for
almost 2 weeks and it's silly to block the sheriff-o-matic critical
path any longer.

This is a script which knows how to walk all the masters in
gatekeeper.json, walk all their builders, grab all their recent builds
find failing steps in those most recent builds, use some (very basic)
"ReasonSplitter" plugins in reason_splitter.py to understand the
"reasons" why those steps failed (failing tests, compile line, etc.)
and then post a stream of these alerts to the data endpoint of your
choice.  Right now the data endpoint is auto-sheriff.appspot.com
but the post format is just a json blob so it could be pointed anywhere.

It also knows how to walk backwards through recent builds and compare
those historical builds to the current failure reasons and thus annotate
the reasons with configuration information about their last_passing config
and their first_failing config.  These config dictionaries are generic but
currently contain the revision information pulled from GOT_REVISION, etc.

The analysis.py module also knows how to merge these alerts into
groups either by reason, or by range.  The data which is posted by default
to auto-sheriff.appspot.com/data includes the raw alert stream
as well as a list of alerts grouped by reason, and then a second grouping
of those reason-grouped alerts by identical config/range information.

These groups include merged_first_fail and merged_last_pass which
represent the min/max of the configs across the group of builders/alerts
which were merged into that group.  So for example if Windows first saw
the failure in a build N which had a smaller chromium revision range than
when Linux saw the failure, these merged configs would have the intersection
of those revision ranges.

All of these parts of this system are designed to be stubs for
possibly more sophisticated replacements, but this has seemed
to be sufficient for now to replace garden-o-matic's data stream
with something generic enough to show failures for chromium's tree
as well as the entire fleet of builders.

Currently this is run on an (auto-updating) cron job on my
personal machine at home, but once this is landed perhaps someone
can help me re-structure it to run on app-engine or on an infra-hosted
GCE VM or build slave.

Currently it's written to use a BuildCache filesystem, but that could be
swapped out for using cloud storage or some other web-acessable
cache of this data.

Earlier I tried to write a appengine-pipeline version of this logic
but failed to fully grok pipelines enough to get it working.

This feed needs to update quickly in order to be useful, right now it
takes a couple minutes to run once the cache is warm and 10+ minutes
to run the first time when the cache is cold.

Originally posted as:
https://codereview.chromium.org/398823008/

TBR=ojan
BUG=399741

Review URL: https://codereview.chromium.org/437593003