Fix builder_alerts bugs with figuring out which steps failed on a builder.
There were two bugs:
1. We incorrectly detected in-progressness of a build by using eta.
The right thing to do is to check if the result is None. There are other
ways to achieve the same thing, but this is what gatekeeper uses,
so may as well do that.
2. If a step has failed in the in-progress build and it failed in the
previous completed build, then we'd include both sets of failures
in the list of failure reasons, which is wrong because we should
only include the failures from the in-progress build.

While I was in this code I removed the "ignored" list of steps from
complete_steps_by_type. It's unused, so no point in maintaining
the code until we need it again in the future.

Also added a couple convenience things to make this easier for
me to hack on locally:
1. --builder-filter to allow only processing certain builder names.
2. If no data_url is provided, write the json out to a local file.

Review URL: https://codereview.chromium.org/472713002