Super-simple HTML view onto /query data
This uses Polymer, which is the way that we've been
writing all of our other new apps for Blink + Infra.

It does some kinda hacky stuff with the url in order
to support all kinds of queries that /query can

It also has very little useful UI, only provides
a table with which to view the results of /query.

This should hopefully aid in debugging and if we
decide we don't like it we can delete it.

BUG=403649
TBR=stip

Review URL: https://codereview.chromium.org/477723002