Add time and state to builder revision data
This patch is intended to make it possible to add alerts to sheriff-o-matic
for offline and hung bots. Prevoiusly, we'd only output the latest
revision information for each bot. This patch also adds a state
(offline, idle, or running) and last update time (last completed run
if offline, or last started step if running).

Review URL: https://codereview.chromium.org/475943002