Remove tests.py from git cl presubmit
Now that test.py is run as part of a try-job as part of
the commit-queue (and git cl try), we no longer need to run
test.py as part of presubmit.

We also run both presubmit.py as well as test.py on the
infra_tester trybot, so currently we're running test.py
twice there, this makes us run it only once.

This brings Infra closer to matching Chromium's setup of
running quick checks as part of presubmit and all
longer testing as part of try-jobs.

BUG=402776

Review URL: https://codereview.chromium.org/483023004