Make builder_alerts work with commit positions.
This gets sheriff-o-matic working again. Down the road we should probably
get sheriff-o-matic/builder_alerts to just operate on git hashes,
but that's a more involved change.

It turns out the parent_ property check in the old code was an outdated
workaround and we don't need it anymore. That's particularly good because
we don't expose _cp versions of the parent properties at the moment.

BUG=401855
R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/501733002