Remove the groupies from the flakiness dashboard
This patch removes groups from the flakiness dashboard, instead showing
all builders.

The two big changes in this mondrianic code review are:
1) The groups feature has been removed entirely.
2) Instead of using {builder, group} as a unique id for builders,
{master, builder} is used. A new class has been added, Builder,
which has a serialization function (.key()) as well as helpers to look
up masters, deserialize, and format the builder name with underscores
to be plaform-neutral.

Review URL: https://codereview.chromium.org/482403003