Add logic to disambiguate conflicting ndb models during test.
This allows multiple ndb-based tests to be loaded into a single interpreter without colliding on model names.

In addition, it detects and raises a TypeError if someone attempts to use a string-based key instead of a class-based one during testing (using a string-based key with this module will break the test in non-obvious ways).

BUG=405649

Review URL: https://codereview.chromium.org/538423002