Make flakiness dashboard links to googlesource work post git migration.
Change links to individual revisions to use crrev.com.

For revision ranges, create a local test-results url that queries
crrev.com for the git hashes and then hits googlesource directly.
This should eventually be part of crrev itself, but this is a
stopgap to get the flakiness dashboard working.

BUG=401856

Review URL: https://codereview.chromium.org/498083002