Fix partytime.
My last change to it made it not work if there were no failures
for a tree because failures[tree] would be undefined. Make it
less brittle, add tests for this code and while here add an
explanatory message that this page means there are no failures.

BUG=407496
NOTRY=true
R=jyasskin@chromium.org

Review URL: https://codereview.chromium.org/530613002

git-svn-id: svn://svn.chromium.org/blink/trunk@181599 bbb929c8-8fbe-4397-9dbb-9b2b20218538