Use existing alert data to avoid walking back in history
For alerts where we've already computed the history, we don't need
to compute it again. This is especially bad when there's a lot of
failures and if they've been happening for a long time.

For alerts where we've already computed the history, we don't need
to compute it again. This is especially bad when there's a lot of
failures and if they've been happening for a long time.

BUG=401351
R=ojan@chromium.org

Review URL: https://codereview.chromium.org/555493002