Fix javascript exception on load_failures.html
This patch fixes an exception on load_failures.html where if the first
bot failed to return results, the loadfailures._failureData.failingBuilders
struct would not be initialized. This bug existed before but was not
hit because the first builder didn't fail to load.

This bug also fixes the counter while loading, and removes the unused
builderToMaster map.

Review URL: https://codereview.chromium.org/553983009