Add framework for Chromium Heartbeat app
Consider this a mini design doc. The idea is to set up an app that expects periodic heartbeat emails from various senders, and alerts a per-sender watchlist if such an email hasn't been received in an expected timeframe.

This app is therefore more generic than gatekeeper-ng-sentry.

Ultimately there should be an interface for configuring this app, so users can just go to chromium-heartbeat.appspot.com, log in, and alter the configs for various heartbeats, or add a brand new one.

BUG=335362

Review URL: https://codereview.chromium.org/560593002