chromium-cq-status: Add individual patchset status timeline view
This change adds a page to chromium-cq-status to display a timeline
of status updates for a given patchset.
This page is primarily HTML/JS and makes use of the /query API, this
means we can migrate it over to another dashboard app relatively easily.

Example of patchset timeline view here:
http://patchset-status.chromium-cq-status.appspot.com/patchset/548223002/60001
This is intended to replace the existing chromium-status page:
https://chromium-status.appspot.com/cq/dstockwell@chromium.org/538503002

BUG=412771

Review URL: https://codereview.chromium.org/560833003