Fix gsubtreed to not require additional tracking refs.
Since we're deterministically creating commit objects, rely on the data which
is currently in the subtree repo (synthesized commit for parent/last tree, and
Mirrored-From footer for position in chromium/src.).

This will cause re-processing of commits which do not affect the subtree mirror
path on every run, but we're hoping that dropping the extra refs will greatly
improve the googlesource replication situation.

R=szager@chromium.org, vadimsh@chromium.org
BUG=407429

Review URL: https://codereview.chromium.org/575623002