Fix test_results auto-recovery from file corruption.
I tried to fix this ages ago, but did it incorrectly because the
tests all use the return value from load_data, but the actual code
uses the data property off the file. Change the tests to use the
data property off the file and fix the auto-recovery to set
self.data to empty string instead of returning it.

Review URL: https://codereview.chromium.org/611203002