Added startup script to infra/
These are python scripts used to create or update a /b checkout.

This down 3 things
1. Set up the /b/.gclient file
2. Seed svn passwords and buildbot passwords
3. Run run_slave.py

Note that an important assumption here is that this is a client script, and it is
the client script's responsibility to make sure that the SVN passwords are
cached correctly.

BUG=
TEST=Ran locally, successfully checkout out a proper slave.DEPS checkout.
TBR=agable

Review URL: https://codereview.chromium.org/654723003