Make builder_alerts give 30 seconds for each bot to fill in failure information.
When a single bot as 1000+ alerts, we don't want to take the time
to fill in all its transition information. Instead, we give it
30 seconds per bot. This gives an upper bound on how long
builder_alerts will take to run instead of having it be O(n*m)
where n is the number of failures and m is how long they've
been failing.

Since we used the previous run's transition information, we
make a little more progress on each run of builder_alerts, so
we end up incrementally computing all the transitions, but in
the meantime we show up to date data to the sheriff.

Review URL: https://codereview.chromium.org/644253004