Clean up twistd pid file before invoking run_slave.py
We know that start.py should be the only thing running run_slave (In fact
the pid file is the PID of chromebuild-startup.py), so if we get to that part of
the code we can safely say that the PID file should be deleted.

This is causing problems on some bots where the twistd processes dies without
cleaning up twistd.pid (Due to a different bug, now fixed).

BUG=406073

Review URL: https://codereview.chromium.org/643773005