Let the lkgr config specify some known revision mappings.
This adds a new config key which lets the config temporarily bypass the
lkgr_finder hash-to-commit-position logic. This is needed in v8 because
the v8 waterfall temporarily ran from a repository which did not have
commit positions, and whose hashes don't appear in the master branch of
the real repository.

R=machenbach@chromium.org
BUG=391686

Review URL: https://codereview.chromium.org/680033002