Make stale master alerts work.
We would crash because we'd call extend with a non-list and
then add the keys in the alert to the list instead of the
alert itself.

Also, master alerts don't have a builder_name, so we need
to keep the gatekeeper apply steps that require builder names.

Also, fixed a silly bug with using + instead of % for
constructing a string. All this affects is the error message
printed out.

Review URL: https://codereview.chromium.org/684703003