chromium-cq-stats: Parameterize analysis on intervals and analyzer classes
This change has no behavioural changes.
This refactors the CQStats analysis logic to be parameterized on interval
lists and analyzer classes. This allows the same analysis logic to be used
for more than updating the stored CQStats to the current time, this
facillitates the ability to rerun particular Analyzers over a specified
time range as an administrative task.

BUG=427448

Review URL: https://codereview.chromium.org/683453002