Defensive coding to treat a M_CLPatchMarked cookie with value "" as 0.
This cookie remembers the user's current position in a list of search results or
among the files in a patchset.  There were reports from users that the cookie was
sometimes set to "" which caused a JS error and prevented the rest of the page
from working.

BUG=426183

Review URL: https://codereview.chromium.org/690533003