Fix gnumbd atomicity issues by doing pushes in-order.
This also will do all pushes serially, per ref, meaning that we'll block
on the pushes instead of queuing them all to the end. With the current
performance levels of gnumbd/googlesource, I don't expect this to be an issue
but we'll need to keep an eye on it.

R=agable@chromium.org,vadimsh@chromium.org,szager@chromium.org
BUG=412731

Review URL: https://codereview.chromium.org/688553002