chromium-cq-status: Use memcache for queries with end timestamps specified
This change adds memcaching to Record and CQStats queries. The cache will
only be used when an end timestamp in the past is specified.
Under this circumstance the results of the query should not change and can
therefore be cached for faster query results.

BUG=433741
R=sergeyberezin@chromium.org

Review URL: https://codereview.chromium.org/725203003