Have the flakiness dashboard minimally support git hashes.
Make it so that things load without errors and so we don't
show links for regression ranges. Down the road, someone
should make links for regression ranges work of course.

Also happens to fix a bug where we'd show NaN for the end
of the regression range for the first item in the list.
Now we just don't show the regression range.

R=pdr@chromium.org

Review URL: https://codereview.chromium.org/741983002