First stab at documenting Endpoints testing.
I compiled the HTML locally and looked at it (this involved disgraceful
alterations to docgen.py), but I'm not sure what specifications I ought to be
cleaving to. I also got 108 warnings, which is a very auspicious number in
many cultures.

I noticed that some of my Python code was automatically syntax-highlighted,
but not all of it. Please let me know if you have insight into that :).

R=pgervais@chromium.org
BUG=

Review URL: https://codereview.chromium.org/810773004