crbuild: added error in buildbucket API responses
Since endpoints does not allow to return 409 http status code and crbuild-buildbot integration needs to understand that lease has expired, return an error reason explicitly. For the sake of consistency, return all errors this way.

All response messages have an error attribute that may be non-null. If
it is, then it contains a reason and a message. Error reason can be
programmed against.

R=vadimsh@chromium.org, sergeyberezin@chromium.org
BUG=

Review URL: https://codereview.chromium.org/814153002