buildbucket: do not check url in start
If a build is started and service.start() is called again, do not check
url, but just overwrite it. It is safe because lease_key is checked, so
there is no race condition.

Context: if a buildbot master is restarted, it may still hold the lease,
but build url is different now. It needs to tell buildbucket that url
has changed. Right now it sends "start" request each time a build
starts.

Also fix run_tests.py

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/852783003