Fix required reviewers saving.
We can't just annotate the required reviewers in the UI with a *, we need
to do it when we're saving too otherwise operations that save the issue
like 'Reopen', or 'Quick LGTM' will drop all the *'s from the reviewers.

This also makes the feature much simpler and fixes the bug where the
loop counter wasn't var'ed leaking into the global scope.

Review URL: https://codereview.chromium.org/868143005