Implement issue searching.
For now it only supports owner, reviewer and cc queries and some basic
sorting. If we need more we can add it, but it's good to start simple
since so many combinations of queries just result in index errors so
the search feature is kind of like rolling dice.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/901383002