Use use_library instead of webapp_django_version.
While webapp_django_version quelled the "You are
using the default Django" message, it triggered
a warning in the log interface.  Switching to
this also quells the local "You are using" message,
and I have a theory it won't trigger the warning
in the logs interface.

R=jrobbins@chromium.org

Review URL: https://codereview.chromium.org/923673002