Avoid a common transaction collision due to saving |surroundings|.
|surroundings| was previously saved as a single page.
All consoles would save their content to a model
instance with that path.  This was triggering transaction
collisions around twice per minute.

Move |surroundings| to be per-console, and save it
always.  Then in the merge process, grab the first
master's surroundings and use that for the subsequent
page data.

R=jrobbins@chromium.org

Review URL: https://codereview.chromium.org/930533002