Remove terrible hack for <cr-patchset-edit-dialog>.
Instead just call stopPropagation() on all taps that come from a
nested dialog. This will prevent the header from collapsing.

This doesn't address the underlying issue that usage of the
issue-refresh event is bad, but it gets rid of the terrible
hack and replaces it with a less gross one.

Review URL: https://codereview.chromium.org/936153005