[Findit] Use absolute url path without hostname:port for the link to the pipeline status page.
Originally, the url used for the link to the pipeline status page includes the hostname and port number, like http://localhost:8080/_ah/pipeline/status?root=bae30af59dbb11e48f83bd7b8a952153&auto=false.

This makes the link inaccessible if the app is running with a different port number locally.
It might cause the same issue if multiple versions of the app are deployed to appengine, because different versions of the app have different hostname.

A solution is to make the url relative to the root of the app. That said, the link above could be converted to: /_ah/pipeline/status?root=bae30af59dbb11e48f83bd7b8a952153&auto=false

BUG=461886

Review URL: https://codereview.chromium.org/957633002