[Findit] Limit size of saved log data of a failed step.
Sometimes, log data is too big to be saved in a single datastore entity with a size limit of 1 MB.
In such case, it is safe to save the ending portion of the log, because failures usually show up at the end.
In compress mode, we could save up to ~4MB of raw log data.

BUG=461916

Review URL: https://codereview.chromium.org/955523002