[Findit] Rename models to make sure they won't conflict with legacy data.
On the prototype appengine https://1-dot-findit-for-me.appspot.com, the legacy data is not compatible with this new version.
But we still wanted to keep the prototype version and its data running for a while.
One easy way to avoid the conflict is to rename the models to make sure they don't have the same name.

BUG=

Review URL: https://codereview.chromium.org/969473002