Switch to the bower version of sugar.
This removes the duplicate copy of sugar and uses the one in third_party
that comes from bower instead. It also allows us to use the minified
version of sugar which reduces the size of the vulcanized app by 178k
unzipped and by 37k zipped.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/964093002