api.trigger: refactoring, made extensible
More an more parameters are being added to trigger_specs that are not
properties. This CL changes api.trigger function to accept trigger specs
instead of properties. A trigger spec may contain properties, as well as
builder_name, buildbot_changes and, in next CL, buildbucket bucket.

Also this change renames "buildbot.changes" to "buildbot_changes",
because a dot cannot be in recipe config schema.

For backward compatibility, if there is "buildername" in a trigger spec,
it is treated in the backward compatible mode.

R=smut@google.com, smut@chromium.org
BUG=449218

Review URL: https://codereview.chromium.org/966993002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294290 0039d316-1c4b-4281-b951-d872f2087c98