Make DiffGroup into a real object.
Instead of just being an array, lets use a real object. This lets
us determine the group type with just a property instead of
needing to inspect the lines. I also simplified the group
logic so we don't even create the group until we decide to
append a line.

I also now mark sections with the class name "section" plus the
a class name of the group type.

R=ojan@chromium.org
BUG=448735

Review URL: https://codereview.chromium.org/976863003