cipd: Add support for symlinks inside CIPD packages.
It is required for packaging of infra.git/ENV virtual python environment, since
it is full of symlinks.

Relative symlinks are packaged as is as long as they point to files inside the
package root. Relative symlinks to files outside of package root are not allowed
and will cause a error during package creation.

Absolute symlinks are allowed. There are two cases:
1. An absolute symlink is pointing to a file inside the package root. Such
symlinks will be converted to relative ones.
2. An absolute symlink is pointing to some unrelated file outside of the package
root. It will be stored as is. This is required to be able to store references
to python executable and stdlib files in the package with python virtualenv.

R=nodir@chromium.org
BUG=

Review URL: https://codereview.chromium.org/970853004