Implement side by side diff
This breaks the DiffBuilder into DiffBuilderBase and then adds two
subclasses DiffBuilderUnified and DiffBuilderSidebyside. It then
refactors some of the surrounding code to allow the diff to be in
either mode while still supporting context line expansion and adding
comments.

For now side-by-side mode is toggled by checking a box in the diff
section. The setting will be persisted in localStorage. Long term
we should come up with a better way to pick which mode you want to
show a file in.

R=ojan@chromium.org
BUG=448735

Review URL: https://codereview.chromium.org/977603004