Use request path instead of query-string to auto-open the revert dialog.
Also, when the user presses the "Revert" button, change the URL to end
with '/revert'.  And, when the dialog closes, remove the '/revert'.
Opening the dialog adds to the browser history so that the user can
also get out of the dialog by using the browser back navigation button.
Closing the dialog removes the dialog state from browser history, so
users stepping back through history will not see the dialog.

The /revert version of the URL also works with the django UI to open
the revert dialog.  If the user switches between the polymer and
django UI's and reloads the same URL or clicks a link that may
end with /revert, then the dialog is shown in either UI.

R=esprehn@chromium.org, esprehn, rmistry
BUG=skia:3458

Review URL: https://codereview.chromium.org/987583003