PatchFile#parseDiff should always return a Promise.
The previous code would either return a Promise or the
actual diff. The caller in loadDiff would also assign
it to this.diff which meant that if the diff was external
the |diff| property would be a Promise while it loaded
and then would turn into the actual diff object later.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/986783002