Fix 'highlight activity' feature.
I accidentally removed the code that allowed turning off the highlights
when I switched to JSON for the inbox. Instead of hacking up
parseInboxData to ignore the recentActivity flag from the server
we just do the real plumbing through all the widgets in this patch. This
also means toggling the checkbox no longer needs to refetch all the
data from the server.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/987773005