Don't screen scrape when saving drafts.
A new JSON API is added at /api/{issueId}/{patchsetId}/{patchId}/draft_message
that accepts a form POST of lineno, text, left and messageId and then
returns the JSON of the newly saved PatchFileMesage (or null if it was deleted).

This removes the flaky screen scraping to find the messageId.

R=jrobbins@chromium.org
BUG=452270

Review URL: https://codereview.chromium.org/990393002