Switch from polymer-ui-collapsible to cr-collapsible.
Stop using the now unsupported <polymer-ui-collapsible> and instead
use a much simpler widget that only supports the functionality that
the codereview tool needs.

In the future we should consider using <core-collapse>, but for
now switching to this widget removes the last of the old
polymer-ui widgets and means all of the UI widgets of the tool
are now self contained.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/997853002