Hide the LGTM and LGTM + CQ buttons for the owner.
Since the current user never changes between app refreshes we store
a property on User instead of a function. This also allows direct
data binding. Doing so makes it easier to show/hide parts of the
UI based on if a given user is the current user.

I also swapped the Revert and Reopen buttons so after clicking the
Close button the Reopen button is where your mouse was.

R=ojan@chromium.org
BUG=466308

Review URL: https://codereview.chromium.org/993353004