Don't rely on the commit property not being on events.
Event handlers get the event object as their argument, this means the
on-tap='{{ send }}' on the Send button was passing the event as the
options and the only reason it didn't commit was because events don't
have a commit property.

Instead use a separate event handler method.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/1001653002