Remove manual embedded language selection.
Highlightjs is capable of picking embedded languages already so all this
code is not needed. It does mean that it's not as easy to add a new
language like 'sky' that contains javascript, but the code
simplification seems worth it.

I think I originally did this because I didn't understand that I was
supposed to keep the syntaxState object around and pass it between
invocations of the highlighter. I then did that, but never removed the
manual language selection code.

This will also fix a bunch of edge cases that my manual selection didn't
work for like doing <script></script><script> which used to leave the
highlighter in markup mode instead of script mode. These cases are all
rare but it does fix them.

I also noticed I had accidentally left a bunch of code behind when I
refactored into the syntax highlighter. This removes that (now dead)
code.

This does seem to produce an empty span when transitioning from a
script or css back into markup which looks like a highlightjs bug.
I don't think the extra few spans at boundaries should be a big
perf issue though.

R=jrobbins@chromium.org

Review URL: https://codereview.chromium.org/998113002