Make CQ bug link text be a call to action.
Instead of having the UI ask questions, this should be an action.
I also used the real url instead of the goo.gl link so users can
tell where they're about to go and made it open in a new tab so
you're not navigated away from the codereview. Finally I removed
the Restrict-View-Google which shouldn't be needed for CQ issues.

Review URL: https://codereview.chromium.org/1085653002