sheriff-o: wraps some app-route contents in <template> tags, also delegates the /health route to a new polymer element so that its nested templates are instantiated correctly when the user visits /health, navigates to some other path and navigates back to /health. Same with the other two modified routes.
This behavior was particularly frustrating to debug because app-router waits for some time to remove the children of the previous route from the dom.  So if you click and navigate back and forth too rapidly, you won't see the bug at all because app-router never gets a chance to remove the dom elements of the previous route.

BUG=

Review URL: https://codereview.chromium.org/1088263002