buildbucket heartbeats: check status
When a build is being heartbeat, check if it is completed before
checking lease. Otherwise, when a build is cancelled on buildbucket, a
buildbot build is cancelled with reason "lease expired" instead of
"build is completed".

Also squash BuildIsCompletedError and InvalidBuildStatusError: the
latter was called only when a build is completed.

R=sheyang@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1127713006