Don't wrap try job masters that have "-" in the name.
When a tryjob is run many times, only the most recent one is shown
with enough width for the label to be visible.  All previous runs are
shown in a colored div that is only 16px wide with overflow hidden.
Any name with a "-" in it caused those colored divs to be multiple
lines tall, which messes up the look of that part of the UI.  This CL
simply limits those divs to the height of one line.

R=esprehn@chromium.org, seanmccullough@chromium.org
BUG=79814

Review URL: https://codereview.chromium.org/1125733007