Add a service_manager service that controls long-running processes.
This is intended to replace sysmon's cron-style launcher and also be
used for mastermon.  It may also be useful for CQ with some more work.

In the future we can extend the service_manager to health check the
services it starts (instead of just making sure they're running), and do
controlled pushes with rollbacks.

BUG=

Review URL: https://codereview.chromium.org/1161513006