sheriff-o: adds a trooper-message banner across the header, if a trooper message is active.
This is useful for letting sheriffs and others know about known outages and such.

demo (read and render the banner): https://91daab4-dot-sheriff-o-matic-staging.appspot.com/chromium
demo (set/deactivate a trooper banner message): https://91daab4-dot-sheriff-o-matic-staging.appspot.com/trooper-msg-form

Note that for the form, you need to be an admin of sheriff-o-matic-*staging* (sent some invites) and you need to be logged in.

/trooper-msg only responds to GET, and it just returns the message if one is active.
/trooer-msg-form is authorized for SoM admins only. Use this form to either set a current banner message, or deactivate it if one is currently active.

The message text is html escaped, but we can probably do something with markdown if that's desireable.

The form is ugly. I can make it nicer if the functionality is satisfactory.  The admins-only authorization is a hack. Do we have a way to ACL GAE apps to say, just chromium-troopers@?

BUG=

Review URL: https://codereview.chromium.org/1149883010