Glyco: added the "pack" command
It is currently only a very sophisticated way of running 'pip wheel' for
packages with a setup.py, that adds the sha1 of the resulting file in its
name.

Use like that:
glyco pack <package directory> -o <where to place the resulting file>

Several packages can be specified on the command-line.

BUG=489420
R=vadimsh@chromium.org

Review URL: https://codereview.chromium.org/1169603004