Work around node.py's race condition if more than one instance runs at a time
Problem: When multiple node.py instances run at the same time for the first time,
the check to see whether or not the installation occured already.  But then
they all race to see who's the first to run shutil.move(), which obviously
fails for everyone other than the first instance.  This CL makes shutil.move()
not fail, since its assumed that if it fails that means someone else already
created an installation.

Another approach is to use an flock, but then it starts to get messy when you
have to keep polling filesystem state to see if another instance finished, or
add timeouts to remove an flock if it was left on the system by a failed attempt,
etc, etc.  This just seemed like a less flaky solution, despite the fact that it
means multiple network requests are spawned.

BUG=
TEST=./node.py & ./node.py & ./node.py & ./node.py & ./node.py

Review URL: https://codereview.chromium.org/1180623002