Update top-level OWNERS in infra.git
It's now the people actually working in Chrome Infra.

The idea is that any one of us should be able to create a new tool or
app in the repo.  If a particular component needs tighter owners, just
'set noparent'.

This is potentially a "political" change (who's got power), so adding
several reviewers in case there are issues.

R=pgervais@chromium.org, agable@chromium.org, abw@chromium.org
BUG=499501

Review URL: https://codereview.chromium.org/1173353004