Remove the migrate_accounts feature.
This feature has not worked in at least 6 months since it's
queues.yaml entry was removed. And, in the CL that removed the entry,
it said that the feature had not worked for all of 2014.
We had one request to migrate accounts for the first time today, and
it was not a request that this feature could have handled anyway.

R=rmistry@google.com, rmistry@chromium.org

Review URL: https://codereview.chromium.org/1177283002