Load ts_mon options from a per-machine config file.
The credentials file and the endpoint are static for a given machine,
and are currently passed through to individual services by puppet or
have to be hardcoded in buildbot recipes
(https://chromereviews.googleplex.com/208137013/).
However generalising this is hard (as seen in
https://chromereviews.googleplex.com/199567013/), and will be even
harder once we move to services started by service_manager where the
service's args aren't controlled by puppet.

BUG=

Review URL: https://codereview.chromium.org/1172273002