Show flaky runs in the flakiness dashboard UI.
This changes the format of results-small.json and results.json to
include all encoded test results for a given run as a string instead
of a single character for just the first failure. The dashboard
labels these runs with the FLAKY color (turqoise) and shows the full
list of statuses on hover.

BUG=279662

Review URL: https://codereview.chromium.org/1182493005