Endpoints for Milo's buildbot json cache
buildbot.py is meant to be used as a local (to the app) broker for acquiring
Buildbot json data.  It acts as a pass-through cache that's smart enough to
consult CBE for data.

This also includes a partial python implementation of https://codereview.chromium.org/1039993004/ (which does things in javascript)

Depends on: https://codereview.chromium.org/1040003002/

BUG=

Review URL: https://codereview.chromium.org/1042763002