[Findit] Use Git Blame to add confidence on Findit result for DEPS rolls.
Use git blame for each file in log to get the exact revision and its commit time. If this time is later than the commit time of old revision and no later than the commit time of the new revision, consider this is a valid hint.

1. If the file in log doesn't exist in the dependency repo, bail out.
2. If the file in log is not changed at all, bail out. (We could use the timestamp of a committed CL to check whether a CL is among the dependency roll.)
3. If the file and line number in log is changed, give a high score 4.
4. If the file is just changed at a different line, give a low score 1.
5. If the file is added, give a high score 5.
6. Without pulling change log for each CL in the rolled range, it seems impossible to detect whether a file was deleted or not.

Currently bail out for v8 repo because of different roll strategy.

BUG=469870

Review URL: https://codereview.chromium.org/1209163003