Add project drop-down menu for chrome infra console app.
R=sergeyberezin@chromium.org
BUG=509895

URL: https://chrome-infra-console-dev.appspot.com/

The dropdown menu on the siderbar allows user choose an available project to display a set of graphs for that project. Once user selects a project, the graphs of selected project would be fetched and replace the current content on the home page.

Review URL: https://codereview.chromium.org/1238623002