[Findit]Accept timezone info when comparing author times.
When Findit suspects some build failures are causeed by a DEPS roll, we need to compare the blame time and the time file is changed to confirm the suspicion.

This change is fixing the logic so we can still handle the author time when it has timezone information.

BUG=509020

Review URL: https://codereview.chromium.org/1247183002