[Findit] Adding logic to determine related file extensions
Files of specific extensions may or may not be related based on extension. Presently, related files are determined using heuristics based primarily on file name, but not extension. This can lead to false positives between files that show up in logs that are actually unrelated to changed files. For example, .h and .c files are likely to be related, but not .isolate and .obj. This change adds logic to associate file extensions with others that are more likely related.

BUG=510528

Review URL: https://codereview.chromium.org/1259153003