[Findit] Fixing logic to normalize object files to account for c, cc, and cpp files.
A corner case for files in the form file.cc.obj is misinterpreted as target.cc.obj, so the logic in _NormalizeObjectFilePath omits 'file' thinking it's the target, then uses 'cc' as the filename, thus returning 'cc.obj'. The correct return result should be 'file.cc.obj'. This change special-cases c, cc, and cpp not to be interpreted as filenames.

BUG=514506

Review URL: https://codereview.chromium.org/1266703003