[Findit] Fix a break in build by 8ec897443393910b0aee245a1a94c0c5ebbc2327.
This is an attempt to fix the break in build https://build.chromium.org/p/chromium.infra/builders/infra-continuous-trusty-32/builds/770.
The root cause might be that the value of the original timestamp is too large for 32bit Linux OS, because the test passed on 64bit Linux OS in the infra waterfall but failed on 32bit ones.

That's probably why the original CL (https://codereview.chromium.org/1261643005/) passed locally and CQ.

TBR=qyearsley

Review URL: https://codereview.chromium.org/1262563005